requirements.txt
fastapi
uvicorn
requests
python-multipart
google-generativeai